feat: Implement product management page with CRUD functionality
- Added Productos page for managing products including add, edit, and delete features.
- Integrated modal for product form input.
- Implemented API calls to fetch, create, update, and delete products.

feat: Create settings page with user profile and options

- Developed Ajustes page for user settings.
- Included profile section with editable fields.
- Added modal for displaying additional information and options.

feat: Add statistics page for sales data visualization

- Created Estadisticas page to display various sales statistics using charts.
- Implemented responsive design for different screen sizes.
- Integrated multiple chart types for data representation.

feat: Build user profile management page

- Developed Perfil page for user profile editing.
- Included image selection from camera or gallery.
- Implemented form validation and update functionality.

feat: Create clients management page with CRUD operations

- Added Clientes page for managing client information.
- Integrated modal for client form input.
- Implemented API calls to fetch, create, update, and delete clients.
