+++ pollosmovil/App.jsx
@@ -1,74 +1,26 @@
 import React, { useState } from "react";
-import { View, StatusBar } from "react-native";
+import { StatusBar } from "react-native";
 import { SafeAreaProvider, SafeAreaView } from "react-native-safe-area-context";
-import {
-  NavigationContainer,
-  createNavigationContainerRef,
-} from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { NavigationContainer } from "@react-navigation/native";
 
-// Pages and Components
-import Login from "./src/pages/Login.jsx";
-import Menu from "./src/components/bottom.navigation.jsx";
-import Dashboard from "./src/pages/Dashboard.jsx";
-import negocios from "./src/pages/business.jsx";
-import estadisticas from "./src/pages/estadisticas.jsx";
-import ajustes from "./src/pages/ajustes.jsx";
-import productos from "./src/pages/productos.jsx";
-import clientes from "./src/pages/clientes.jsx";
-import Perfil from "./src/pages/Perfil.jsx";
-import Modal from "./src/components/Modal.componet.jsx";
-
-const Stack = createNativeStackNavigator();
-const navigationRef = createNavigationContainerRef();
-const MENU_HEIGHT = 62;
+import AuthNavigator from "./src/navigation/AuthNavigator.jsx";
+import AppNavigator from "./src/navigation/AppNavigator.jsx";
 
 export default function App() {
-  const [currentRoute, setCurrentRoute] = useState(null);
   const [isLoggedIn, setIsLoggedIn] = useState(false);
 
   return (
     <SafeAreaProvider>
-      <StatusBar
-        translucent={false}
-        backgroundColor="#ffffff"
-        barStyle="dark-content"
-      />
+      <StatusBar backgroundColor="#fff" barStyle="dark-content" />
+      <SafeAreaView style={{ flex: 1 }}>
+        <NavigationContainer>
+          {!isLoggedIn ? (
+            <AuthNavigator onLoginSuccess={() => setIsLoggedIn(true)} /> 
+          ) : (
+            <AppNavigator />
+          )}
+        </NavigationContainer>
 
-      <SafeAreaView style={{ flex: 1, backgroundColor: "#fff" }}>
-
-        {!isLoggedIn ? (
-          <Login onLoginSuccess={() => setIsLoggedIn(true)} />
-        ) : (
-          <NavigationContainer
-            ref={navigationRef}
-            onReady={() =>
-              setCurrentRoute(navigationRef.getCurrentRoute()?.name)
-            }
-            onStateChange={() =>
-              setCurrentRoute(navigationRef.getCurrentRoute()?.name)
-            }
-          >
-            <View style={{ flex: 1, paddingBottom: MENU_HEIGHT }}>
-              <Stack.Navigator screenOptions={{ headerShown: false }}>
-                <Stack.Screen name="app" component={Dashboard} />
-                <Stack.Screen name="negocios" component={negocios} />
-                <Stack.Screen name="estadisticas" component={estadisticas} />
-                <Stack.Screen name="ajustes" component={ajustes} />
-                <Stack.Screen name="productos" component={productos} />
-                <Stack.Screen name="clientes" component={clientes} />
-                <Stack.Screen name="Perfil" component={Perfil} />
-                <Stack.Screen name="Modal" component={Modal} />
-              </Stack.Navigator>
-
-              <Menu
-                navigationRef={navigationRef}
-                currentRouteName={currentRoute}
-                height={MENU_HEIGHT}
-              />
-            </View>
-          </NavigationContainer>
-        )}
       </SafeAreaView>
     </SafeAreaProvider>
   );
