+++ pollosmovil/src/components/common/Graficas.componet.jsx
