+++ pollosmovil/src/pages/Users/clientes.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
 import { View, Text, StyleSheet, Image, TouchableOpacity, TextInput, ScrollView, Alert } from "react-native";
 import Icon from "react-native-vector-icons/MaterialCommunityIcons";
-import HeaderPrincipal from "../components/header";
-import Modal from "../components/Modal.componet";
-import { getClients, createClient, updateClient, deleteClient, } from "../Hook/Api/clientApi";
+import HeaderPrincipal from "../../components/layout/header";
+import Modal from "../../components/common/Modal.componet";
+import { getClients, createClient, updateClient, deleteClient, } from "../../Hook/Api/clientApi";
 
 const Clients = () => {
   const [clients, setClients] = useState([]);
